requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1; python_version < '3.9'
boto3==1.26.79
botocore==1.29.79
crispy-bootstrap5==0.7
Django>=3.2
django-crispy-forms==2.0
django-fontawesome-5==1.0.18
django-icons==22.1
django-storages==1.13.2
fontawesomefree==6.3.0
jmespath==1.0.1
pillow==9.4.0
python-dateutil==2.8.2
psycopg2-binary>=2.8
python-decouple==3.7
python-dotenv>=1.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uWSGI>=2.0
urllib3==1.26.14